Icinga DB: merge SendStatusUpdate into UpdateState
On state changes, we sometimes want to only update Redis and sometimes want to
update both Redis and SQL. Before this commit, updating the Redis
icinga:*:state hashes (this is used when icingadb-web reads from Redis) and
generating a runtime update (this is what writes through changes to SQL) was
done by separate functions and it was up to the caller to do the right thing.

However, this allowed also to just generate a runtime update which makes no
sense as these should always reflect changes done to the Redis hashes.
Therefore, merge both functions into one that always updates the hashes and
optionally also generates a runtime update. The runtime update will be skipped
in two cases: the initial sync (the change will be written to SQL by a full
update anyways) and minor state updates (i.e. just the plugin output chaning
which should not be written through to the SQL database to save operations
there).